+++ modules/generator.cpp
@@ -1,21 +1,21 @@
 #include "generator.hpp"
 
 
-Generator::Generator(std::string name, float hull, float power_output, float fuel_consumption): 
+Generator::Generator(std::string name, double hull, double power_output, double fuel_consumption): 
 Module(name, hull), 
 max_power_output(power_output), 
 max_fuel_consumption(fuel_consumption)
 {
 }
 
-void Generator::simulate(float delta_time, Ship * ship)
+void Generator::simulate(double delta_time, Ship * ship)
 {
     if (this->enabled == false) { return; }
     
     //try to add max power
-    float fuel_available = ship->getFuel();
-    float energy_output = max_power_output*delta_time;
-    float fuel_needed = max_fuel_consumption*delta_time;
+    double fuel_available = ship->getFuel();
+    double energy_output = max_power_output*delta_time;
+    double fuel_needed = max_fuel_consumption*delta_time;
     
     //not enough fuel
     if (fuel_needed > fuel_available)
@@ -25,7 +25,7 @@ void Generator::simulate(float delta_time, Ship * ship)
     }
     
     //try to save energy
-    float real_energy_output = ship->addEnergy(energy_output);
+    double real_energy_output = ship->addEnergy(energy_output);
     std::cout << "real_energy_output " << real_energy_output << std::endl;
     
     //not enough capacity
